+++ py/desisurvey/scripts/surveymovie.py
@@ -67,6 +67,9 @@ def parse(options=None):
     parser.add_argument(
         '--output-path', default=None, metavar='PATH',
         help='output path where output files should be written')
+    parser.add_argument(
+        '--config-file', default='config.yaml', metavar='CONFIG',
+        help='input configuration file')
 
     if options is None:
         args = parser.parse_args()
@@ -75,7 +78,7 @@ def parse(options=None):
 
     # Validate start/stop date args and covert to datetime objects.
     # Unspecified values are taken from our config.
-    config = desisurvey.config.Configuration()
+    config = desisurvey.config.Configuration(args.config_file)
     if args.start is None:
         args.start = config.first_day()
     else: